loaders/requirements.txt
azure-common==1.1.28
azure-core==1.27.0
azure-functions==1.14.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
pandas==2.0.1
psycopg2==2.9.6
psycopg2_binary==2.9.6
pytest==7.3.1
python-dotenv==1.0.0
